+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JZ0ESKT6FAWCQ3NP99EFD452"
+        "01JZ13NS9EE31DY2VXR0X0N5PP"
     )
     r"""The id of the resource"""