+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -59,5 +59,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-06-30T13:00:29.711Z")
+    created: Optional[datetime] = parse_datetime("2025-06-30T19:05:24.368Z")
     r"""The date and time the resource was created"""