+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JZ0ESKT4N6MMN421FWVBZ65E"
+        "01JZ13NS9D8H26MDZZ167H0Z7G"
     )
     r"""The id of the resource"""