+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -572,5 +572,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-06-30T13:00:26.696Z")
+    updated: Optional[datetime] = parse_datetime("2025-06-30T19:05:21.426Z")
     r"""The date and time the resource was last updated"""