+++ packages/orq-rc/RELEASES.md
@@ -1889,3 +1889,13 @@ Based on:
 - [python v3.9.0-rc.10] packages/orq-rc
 ### Releases
 - [PyPI v3.9.0-rc.10] https://pypi.org/project/orq-ai-sdk/3.9.0-rc.10 - packages/orq-rc
+
+## 2025-06-30 19:06:58
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.572.0 (2.643.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.9.0-rc.11] packages/orq-rc
+### Releases
+- [PyPI v3.9.0-rc.11] https://pypi.org/project/orq-ai-sdk/3.9.0-rc.11 - packages/orq-rc